Fix deprecation
Rename plugin with prefix _

Fix validate-modules bug introduced in 2.18 for ansible.builtin plugins
