+++ test/lib/ansible_test/_util/controller/sanity/validate-modules/validate_modules/schema.py
@@ -783,10 +783,10 @@ def deprecation_schema(for_collection):
         Required('why'): doc_string,
         'alternative': doc_string,
         'alternatives': doc_string,
+        Required('removed_from_collection'): collection_name,
     }
 
-    if for_collection:
-        main_fields.update({Required('removed_from_collection'): collection_name, 'removed': Any(True)})
+    main_fields.update({'removed': Any(True)})
 
     date_schema = {
         Required('removed_at_date'): date(),
